Move helper methods to ProtocolConversions
Purely mechanical, except for making one method public
